WARNS=4 fixes.
OK'ed by:	des

+++ contrib/libpam/libpamc/include/security/pam_client.h
@@ -1,5 +1,6 @@
 /*
  * $Id: pam_client.h,v 1.4 2001/01/20 22:29:47 agmorgan Exp $
+ * $FreeBSD$
  *
  * Copyright (c) 1999 Andrew G. Morgan <morgan@linux.kernel.org>
  *
@@ -136,7 +137,7 @@ do {                                                                       \
 
 #define PAM_BP_FILL(prmpt, offset, length, data)                           \
 do {                                                                       \
-    int bp_length;                                                         \
+    size_t bp_length;                                                         \
     __u8 *prompt = (__u8 *) (prmpt);                                       \
     bp_length = PAM_BP_LENGTH(prompt);                                     \
     if (bp_length < ((length)+(offset))) {                                 \
@@ -147,7 +148,7 @@ do {                                                                       \
 
 #define PAM_BP_EXTRACT(prmpt, offset, length, data)                        \
 do {                                                                       \
-    int __bp_length;                                                       \
+    size_t __bp_length;                                                       \
     const __u8 *__prompt = (const __u8 *) (prmpt);                         \
     __bp_length = PAM_BP_LENGTH(__prompt);                                 \
     if (((offset) < 0) || (__bp_length < ((length)+(offset)))              \
